+++ satella/coding/decorators/__init__.py
@@ -4,11 +4,11 @@ from .arguments import auto_adapt_to_methods, attach_arguments, for_argument, \
     cached_property
 from .decorators import wraps, chain_functions, has_keys, short_none, memoize, return_as_list, \
     default_return, cache_memoize, call_method_on_exception
-from .flow_control import loop_while, queue_get
+from .flow_control import loop_while, queue_get, repeat_forever
 from .preconditions import postcondition, precondition
 from .retry_dec import retry
 
-__all__ = ['retry', 'transform_result', 'transform_arguments',
+__all__ = ['retry', 'transform_result', 'transform_arguments', 'repeat_forever',
            'execute_before', 'postcondition', 'precondition', 'wraps', 'queue_get',
            'chain_functions', 'has_keys', 'short_none', 'auto_adapt_to_methods',
            'attach_arguments', 'for_argument', 'loop_while', 'memoize',