+++ satella/instrumentation/cpu_time/__init__.py
@@ -1,6 +1,6 @@
-from .collector import calculate_occupancy_factor, sleep_cpu_aware, CPUTimeManager
+from .collector import calculate_occupancy_factor, sleep_cpu_aware, CPUTimeManager, get_own_cpu_usage, pCPUtimes
 from .concurrency import CPUTimeAwareIntervalTerminableThread
 
-__all__ = ['calculate_occupancy_factor', 'sleep_cpu_aware',
+__all__ = ['calculate_occupancy_factor', 'sleep_cpu_aware', 'pCPUtimes', 'get_own_cpu_usage',
            'CPUTimeAwareIntervalTerminableThread',
            'CPUTimeManager']