+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # v2.23.5
 
+The software
+------------
+
 * added AutoflushFile.seek()
 * added AutoflushFile.truncate()
+* added CPUTimeManager.set_refresh_each()
+* added satella.instrumentation.cpu_time.get_own_cpu_time()
+
+Build process
+-------------
+
+* added pylint